+++ back-end/src/main/java/com/hitpot/web/controller/req/WithdrawForm.java
@@ -8,5 +8,5 @@ import lombok.Data;
 @ApiModel("提现表单")
 public class WithdrawForm {
     @ApiModelProperty("提现pot数量")
-    private Long amountPot;
+    private double amountPot;
 }
